requirements.txt
openai>=1.0.0
requests
beautifulsoup4>=4.12
aiogram>=3.0
aiohttp>=3.8
python-dotenv>=1.0.0
duckduckgo-search>=6.0
fastapi>=0.100
uvicorn>=0.22
jinja2>=3.0
python-multipart>=0.0.6
pyttsx3>=2.90
sounddevice>=0.4.6
soundfile>=1.0.4
numpy>=1.20
